Bias generation has been reconfigured to reduce voltage droop under load
Changes:

10uH -> 15 uH inductor. 15uH gave higher current than either 10 or 22uH.

Larger boost capacitor, and change footprint to 1206 so that capacitance is more stable under bias. This stabilizes output at higher current.

Feedback nowconnects directly to the BIAS and GND pins, thus avoiding the voltage drop across the RC filter. In addition, one of the RC filters is now split across the bias and TIA boards so that the feedback can be taken after the resistor.

More decoupling on boost converter.

Added RC filter on input to LT3571 as the PSRR seems quite low.

New programming header on backside.  Note that I used the uncommon foot print, so finding headers for this is annoying.  I should probably switch to the mirrored one which is what 99% of headers use.