Minor changes to power supply
Changes:

Add LC filter on output of the negative regulator. This is almost certainly not needed, but it further rejects the 2MHz inverting regulator switch frequency.  Replace inductor with a jumper if you don't want to use it.

Clean up traces and layout of existing components.

Widen notch so that SMB cables can fit through more easily.

Adjust negative rail voltage to be -4.5v instead of -3.3v to increase dynamic range.

Move connector so there is a bit more room to fit in lens tube.