Standardize capacitor sizes and update BOM
Previously there was a mix of 2.2uF and 1.0uF filter capacitors.  Either size will work, so it makes little sense to have both sizes.  Switch everything to 2.2uF and make a note that smaller values are acceptable.

Also, add the voltages to the BOM to make clear that the high voltage side must have 100V rated capacitors.