Check in (tested) alternative psu
Can use easier to get parts for the charge pump and positive regulator.

Specs are actually better than the original design but the difference is insignificant.  We bumped the negative voltage rail up a little (-4.6V) due to better parts.

Can use the existing TPS79333 (which is hard to find) or the tps7a20 which has better specs and is pin compatible.  Can optionally use the 4.2 or 4.5v version as well to extend dynamic range, although this won't make a difference with the current bias generator and s14420 due to current limits.  Maybe help if used with other SiPMs.