Make sure the LT836X output voltage is set
Previously on power up it could end up set to incorrect values.  We now initialize it to about 52.8V, which is high enough to cover the entire 0 to 4095 gain range, with 4095 being ~52.4V.  Note that the RT9072 has a drop out voltage of 250mV, so the boost converter must always be set at least 250mV higher than the output voltage or the regulator will not function.