Move the old bias generator out of the repo base
I'll leave it since it might still be useful for some people and its what the Scientific Reports manuscript used, but it is essentially obsolete and fairly hard to assemble.  The LT8362 converter is much easier to make and can supply a lot more power so it should be the default choice.